% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{applyDrift}
\alias{applyDrift}
\title{Apply all drift corrections from a list of corrections}
\usage{
applyDrift(timeSeries, corrections)
}
\arguments{
\item{timeSeries}{a time series with all fouling corrections applied to a column
called foulingCorrected}

\item{corrections}{a list of USGS multi-point corrections from \code{getCorrections}}
}
\value{
a numeric vector of the combined effect of all the fouling drift in the
list
}
\description{
Apply each Set 2 correction from a list of multiPointCorrections to a
time series that has had fouling corrections applied to a column foulingCorrected. 
(Should not be used directly, only from within corrApply)
}
