% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{corrApply}
\alias{corrApply}
\title{Create a data frame of raw time series data with corrections annotated}
\usage{
corrApply(tsID, startDate, endDate)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{startDate}{the start date of interest as a string in the form YYYY-MM-DD}

\item{endDate}{the end date of interest as a string in the form YYYY-MM-DD}
}
\description{
Create a data frame of raw time series data with corrections annotated
}
