% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{findDisapproval}
\alias{findDisapproval}
\title{Find periods that have been set back from approved}
\usage{
findDisapproval(approvalList)
}
\arguments{
\item{approvalList}{a data frame of approval transaction sfrom getApprovalList()}
}
\description{
Given a list of approval transactions (from \code{getApprovalList}), give any 
periods that had been set to approved, but were later set to a lower status, 
whether working or in review
}
