% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{genericAQ}
\alias{genericAQ}
\title{Base function to get data or metadata from the AQ Publish API}
\usage{
genericAQ(resource, parameters, values, host = "ts-api.nwis.usgs.gov")
}
\arguments{
\item{resource}{the resource to retrieve - such as raw data (GetTimeSeriesRawData),
or corrections (GetCorrectionList). See AQ API documentation for details on available
resources}

\item{parameters}{a character vector the names of the parameters to query by}

\item{values}{the values corresponding to each parameter}

\item{host}{the host to use - could be a test server for development, or the production
server for running real queries}
}
\value{
the response from the query
}
\description{
Given a service request and a set of query parameters and values, return the 
output from the API
}
