% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getApprovalList}
\alias{getApprovalList}
\title{Get a list of approval transactions}
\usage{
getApprovalList(timeSeriesID, start, end)
}
\arguments{
\item{timeSeriesID}{the time series unique identifier}

\item{start}{the starting date of interest, as a string in the form YYYY-MM-DD}

\item{end}{the ending date of interest, as a string in the format YYYY-MM-DD}
}
\value{
a data frame of approval transactions
}
\description{
Get a data frame with details about when approval levels were changed for a given time series
in a particular date range
}
