% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getCorrections}
\alias{getCorrections}
\title{Get a list of USGS Multi-point corrections}
\usage{
getCorrections(tsID, start, end)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{start}{the start date of interest as a string in the form YYYY-MM-DD}

\item{end}{the end date of interest as a string in the form YYYY-MM-DD}
}
\value{
a list of multiPointCorrection objects
}
\description{
Get a list of USGS Multi-point corrections
}
