% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getGapTolerance}
\alias{getGapTolerance}
\title{Get gap tolerances from AQUARIUS}
\usage{
getGapTolerance(tsID, start, end)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{start}{the start date of interest as a string in the form YYYY-MM-DD}

\item{end}{the end date of interest as a string in the form YYYY-MM-DD}
}
\description{
Get gap tolerances from AQUARIUS
}
