% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_flows.R
\name{getSampleQ}
\alias{getSampleQ}
\title{Get a table of sample times, flows and gage heights}
\usage{
getSampleQ(site, start, end, maxDiff = 4, method = c("interpolate",
  "closest"), tz = "UTC")
}
\arguments{
\item{site}{USGS site number}

\item{start}{the start date of interest, as a string in the form YYYY-MM-DD}

\item{end}{the start date of interest, as a string in the form YYYY-MM-DD}

\item{maxDiff}{the maximum gap, in hours, in the continuous data that 
will be used for merging. Any gaps greater than this will show NA for the
continuous data.}

\item{method}{the method to use for merging the continuous data. If "interpolate",
the continuous value will be interpolated for the sample time between the time
series points immediately preceding and following. If "closest", the continuous
point nearest the sample time will be used.}

\item{tz}{the time zone for output, e.g. "America/Chicago", "America/New_York", 
see \code{OlsonNames()} for more details}
}
\value{
a data frame of QW sample times, streamflow, and discharge merged with 
continuous streamflow and discharge for comparison.
}
\description{
The table includes QW sample times, as well as any associated gage heights
and discharge values associated with the sample. Discharge and gage height
are interpolated form the relevant continuous time series for comparison.
}
