% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tools.R
\name{getSingleTS}
\alias{getSingleTS}
\title{Get a single time series}
\usage{
getSingleTS(tsID, start, end, name)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{start}{the starting date of interest as a string in the form YYYY-MM-DD}

\item{end}{the ending date of itnerest as a string in the form YYYY-MM-DD}

\item{name}{the name to use in the data column}
}
\description{
A wrapper for getCorrectedTimeSeries, but with the option to 
rename the data column in the resulting data
}
