% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tools.R
\name{getSingleTSSplit}
\alias{getSingleTSSplit}
\title{Get a single time series, separating the query into 3 year chunks}
\usage{
getSingleTSSplit(tsID, start, end, name)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{start}{the starting date of interest as a string in the form YYYY-MM-DD}

\item{end}{the ending date of itnerest as a string in the form YYYY-MM-DD}

\item{name}{the name to use in the data column}
}
\description{
Get a single time series, separating the query into 3 year chunks
}
