% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_tools.R
\name{getTimeSeriesConnect}
\alias{getTimeSeriesConnect}
\title{Authenticate API connection and get multiple time series joined into one data frame}
\usage{
getTimeSeriesConnect(tsID, start, end, names, time_zone,
  interval = c("All", "Hourly", "30-minute", "15-minute", "5-minute"),
  id, pw)
}
\arguments{
\item{tsID}{the time series unique identifiers for each series}

\item{start}{the starting date of interest as a string in the form YYYY-MM-DD}

\item{end}{the ending date of itnerest as a string in the form YYYY-MM-DD}

\item{names}{the names to use for each series, corresponding to tsID}

\item{time_zone}{the time zone to use for output}

\item{id}{the AQUARIUS API username to authenticate with}

\item{pw}{the AQUARIUS API encrypted password to authenticate with}
}
\description{
For use in asynchronous programming for the shiny application
}
