% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getTimeSeriesIDs}
\alias{getTimeSeriesIDs}
\title{Get time series descriptions, including unique identifiers}
\usage{
getTimeSeriesIDs(location, parameter)
}
\arguments{
\item{location}{the location identifier}

\item{parameter}{the parameter name}
}
\value{
a data frame of time series descriptions matching the given location and parameter,
including a column giving the unique identifier for each time series
}
\description{
Get a list of time series descriptions based on the location identifier and parameter name
}
