% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getToken}
\alias{getToken}
\title{Authenticate to the AQ API}
\usage{
getToken(id, pw)
}
\arguments{
\item{id}{the username to use for authentication}

\item{pw}{the encrypted password to use for authentication}
}
\value{
the response from the authentication
}
\description{
Use the username and encrypted password to authenticate to the AQUARIUS API
}
