% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_flows.R
\name{intQGHT}
\alias{intQGHT}
\title{Merge sample times and continuous data by interpolation}
\usage{
intQGHT(sample_data, cont_data, maxDiff = 4)
}
\arguments{
\item{sample_data}{sample times from \code{getSampleTimes}}

\item{cont_data}{continuous streamflow and gage height data from \code{getQGHT}}

\item{maxDiff}{the maximum gap, in hours, in the continuous data that 
will be used for merging. Any gaps greater than this will show NA for the
continuous data}
}
\description{
Merge sample times and continuous data by interpolation
}
