% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{isOverlap}
\alias{isOverlap}
\title{Determine whether two intervals overlap}
\usage{
isOverlap(start1, end1, start2, end2)
}
\arguments{
\item{start1}{the start of the first interval}

\item{end1}{the end of the first interval}

\item{start2}{the start of the second interval}

\item{end2}{the end of the second interval}
}
\value{
a logical indicating whether the two intervals overlap
}
\description{
Determine whether two intervals overlap
}
