% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{makeTable}
\alias{makeTable}
\title{Make a complete table of correction data and grades}
\usage{
makeTable(tsID, start, end, parm)
}
\arguments{
\item{tsID}{the unique time series identifier}

\item{start}{the start date of interest as a string in the form YYYY-MM-DD}

\item{end}{the end date of interest as a string in the form YYYY-MM-DD}

\item{parm}{the parameter name}
}
\description{
Make a complete table of correction data and grades
}
