% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{makeTableConnect}
\alias{makeTableConnect}
\title{Wrapper for makeTable, with internal API authentication}
\usage{
makeTableConnect(tsID, start, end, parm, id, pw)
}
\arguments{
\item{tsID}{the unique time series identifier}

\item{start}{the start date of interest as a string in the form YYYY-MM-DD}

\item{end}{the end date of interest as a string in the form YYYY-MM-DD}

\item{parm}{the parameter name}

\item{id}{the username to use for authentication}

\item{pw}{the encrypted passowrd to use for authentication}
}
\description{
Used for asynchronous programming in the shiny application.
}
