% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\docType{class}
\name{multiPointCorrection-class}
\alias{multiPointCorrection-class}
\alias{multiPointCorrection}
\title{An S4 class to represent a USGS multipoint correction}
\description{
An S4 class to represent a USGS multipoint correction
}
\section{Slots}{

\describe{
\item{\code{startTime}}{the beginning time of the correction, a POSIXct object.}

\item{\code{endTime}}{the ending time of the correction, a POSIXct object.}

\item{\code{set}}{the set the correction was applied in, an integer - either 1 (fouling), 
2 (drift), or 3 (other).}

\item{\code{startValues}}{the values to which the starting offsets are applied.
Must be the same length as startOffsets.}

\item{\code{startOffsets}}{the starting offsets of the correction. Must be the same
length as startValues.}

\item{\code{endValues}}{the values to which the ending offsets are applied.
Must be the same length as endOffsets.}

\item{\code{endOffsets}}{the ending offsets of the correction. Must be the same
length as endValues.}
}}

