% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{recordCompleteness}
\alias{recordCompleteness}
\title{Information for evaluating record completeness}
\usage{
recordCompleteness(datetimes, start = "auto", end = "auto",
  freq = "auto")
}
\arguments{
\item{datetimes}{a vector of datetimes as POSIXct}

\item{start}{the start date of interest, as POSIXct. If "auto", the first date
in datetimes will be used}

\item{end}{the end date of itnerest, as POSIXct. If "auto", the last date
in datetimes will be used}

\item{freq}{the time series frequency, in minutes. If "auto", the most common
frequency will be used}
}
\value{
a data frame with information about record completeness
}
\description{
Information for evaluating record completeness
}
