% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{retryToken}
\alias{retryToken}
\title{Retry authentication}
\usage{
retryToken(id, pw)
}
\arguments{
\item{id}{the username to use for authentication}

\item{pw}{the encrypted password to use for authentication}
}
\value{
the response from getToken from the succesful authentication or the
failed reponse from the final try
}
\description{
Retry authentication using the username and password until it works, or a maximum
of 5 times
}
