% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{summarizeGaps}
\alias{summarizeGaps}
\title{Summarize and tabulate the results of findGaps}
\usage{
summarizeGaps(gapTest, gapTol)
}
\arguments{
\item{gapTest}{the results of \code{findGaps}}

\item{gapTol}{the gap tolerance, in minutes or a data frame of gap
tolerances from \code{getGapTolerance}}
}
\value{
a data frame summary of gaps
}
\description{
Give a data frame of information about how much of a time period is classified
as a gap in the data record
}
