% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{summarizeGrades}
\alias{summarizeGrades}
\title{Summarize grades from the complete table}
\usage{
summarizeGrades(dataTable)
}
\arguments{
\item{dataTable}{a table output from \code{makeTable}}
}
\description{
Summarize grades from the complete table
}
