% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{testToken}
\alias{testToken}
\title{Test to see if the API connection is working}
\usage{
testToken()
}
\value{
a logical indicating whether the API connection is working, and still active
}
\description{
Test to see if the API connection is working
}
