% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{wagnerGrade}
\alias{wagnerGrade}
\title{Apply grades based on TM-1D3}
\usage{
wagnerGrade(parameter, raw, percent, numerical)
}
\arguments{
\item{parameter}{the parameter name}

\item{raw}{the raw time series data}

\item{percent}{the total percent correction}

\item{numerical}{the total numerical correction}
}
\value{
a character vector of grades, one of "Excellent", "Good", 
"Fair", "Poor", or "Consider Deletion"
}
\description{
Apply grades based on TM-1D3
}
