% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{whatOverlap}
\alias{whatOverlap}
\title{Find the overlap of two overlapping intervals}
\usage{
whatOverlap(start1, end1, start2, end2)
}
\arguments{
\item{start1}{the start of the first interval}

\item{end1}{the end of the first interval}

\item{start2}{the start of the second interval}

\item{end2}{the end of the second interval}
}
\value{
a vector of length 2 giving the start and end of the overlap
between the two intevals
}
\description{
Find the overlap of two overlapping intervals
}
