% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{applyCorrection}
\alias{applyCorrection}
\title{Apply a correction to a time series}
\usage{
applyCorrection(datetime, raw, correction)
}
\arguments{
\item{datetime}{a vector of datetimes, as POSIXct}

\item{raw}{a numeric vector of raw time series data - should be the same length
as datetime}

\item{correction}{a multiPointCorrection object}
}
\value{
a numeric data of the value of the correction at each point
}
\description{
Apply a correction to a time series
}
