% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{corrApply}
\alias{corrApply}
\title{Create a data frame of time series data with details on corrections}
\usage{
corrApply(tsID, start, end)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{startDate}{the start date of interest as a string in the form YYYY-MM-DD}

\item{endDate}{the end date of interest as a string in the form YYYY-MM-DD}
}
\value{
a data frame of time series data. The value of all fouling
and drift corrections is shown for each point in the time series.
}
\description{
Create a data frame of time series data with details on corrections
}
