% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections_grades_gaps.R
\name{findGaps}
\alias{findGaps}
\title{Find gaps in a time series record}
\usage{
findGaps(datetimes, gapTol = 120)
}
\arguments{
\item{datetimes}{a vector of datetimes}

\item{gapTol}{the gap tolerance, in minutes or a data frame of gap
tolerances from \code{getGapTolerance}}
}
\value{
a data frame of gaps
}
\description{
Find gaps in a time series record
}
