% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getAvailableTimeSeries}
\alias{getAvailableTimeSeries}
\title{Get all available time series for a location and date range}
\usage{
getAvailableTimeSeries(location, start, end, publish)
}
\arguments{
\item{location}{the location identifier}

\item{start}{the start date (or datetime) as a string in a standard 
format, or a POSIXct object}

\item{end}{the end date (or datetime) as a string in a standard 
format, or a POSIXct}
}
\description{
Get all available time series for a location and date range
}
