% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{getCorrectedData}
\alias{getCorrectedData}
\title{Get corrected data for the time series data}
\usage{
getCorrectedData(tsID, start, end)
}
\arguments{
\item{tsID}{the time series unique identifier}

\item{start}{the starting date of interest as a string in the form YYYY-MM-DD}

\item{end}{the ending date of itnerest as a string in the form YYYY-MM-DD}
}
\value{
a data frame of corrected data for the time series
}
\description{
Get time series data from AQUARIUS with all corrections applied
}
