% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_flows.R
\name{getSampleTimes}
\alias{getSampleTimes}
\title{Find QW sample times, with streamflow and gage height, if available}
\usage{
getSampleTimes(site, start, end, tz = "UTC")
}
\arguments{
\item{site}{USGS site number}

\item{start}{the start date of interest, as a string in the form YYYY-MM-DD}

\item{end}{the start date of interest, as a string in the form YYYY-MM-DD}

\item{tz}{the time zone for output, e.g. "America/Chicago", "America/New_York", 
see \code{OlsonNames()} for more details}
}
\description{
Find QW sample times, with streamflow and gage height, if available
}
